requirements.txt
streamlit==1.53.0
python-dotenv==1.2.1
pypdf==6.6.0
spacy==3.8.11
rapidfuzz==3.14.3
langchain==1.2.6
langchain-community==0.4.1
langchain-openai==1.1.7
chromadb==1.4.1
pydantic==2.12.5
pytest==9.0.2